+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [current]
+### Fixed
+- Post agent will now interpret server responses as UTF-8 with an option
+  to override.
 
 ## [0.9.14.2] - 2021-04-18
 ### Fixed